Add node-lite image and decouple node config from cli package
Move the shared node configuration and context classes into vantage6-common
so the node package no longer depends on the top-level vantage6 CLI package.
This drops a large amount of CLI-only dependency baggage from the node image
while keeping the existing CLI import path working through a temporary shim.

Add a dedicated node-lite Dockerfile and local release workflow for building a
smaller node-only image, and make the node package declare the runtime modules
it actually needs in that slimmer image. Keep interactive CLI helpers lazy so
non-interactive node startup does not pull them in unnecessarily.

Lite was briefly named "MDW", but evetually this image (or some form of
it) might be worthy of suggesting upstream.
