+++ .github/workflows/release-node-lite.yml
@@ -1,4 +1,4 @@
-name: Release Node MDW Image
+name: Release Node Lite Image
 
 on:
   push:
@@ -8,7 +8,7 @@ on:
 permissions: {}
 
 jobs:
-  release-node-image:
+  release-node-lite-image:
     runs-on: ubuntu-24.04
     timeout-minutes: 30
     permissions:
@@ -17,7 +17,7 @@ jobs:
       attestations: write
       id-token: write
     env:
-      IMAGE_REPOSITORY: ghcr.io/${{ github.repository }}/infrastructure/node-mdw
+      IMAGE_REPOSITORY: ghcr.io/${{ github.repository }}/infrastructure/node-lite
 
     steps:
       - name: Checkout repository
@@ -69,7 +69,7 @@ jobs:
         uses: docker/build-push-action@bcafcacb16a39f128d818304e6c9c0c18556b85f
         with:
           context: .
-          file: docker/node-mdw.Dockerfile
+          file: docker/node-lite.Dockerfile
           platforms: linux/amd64
           push: true
           tags: ${{ steps.meta.outputs.image_tag }}
