+++ docker/node-lite.Dockerfile
@@ -1,10 +1,10 @@
-# Dockerfile for the MDW node image
+# Dockerfile for the node-lite image
 #
-# MDW uses a more minimal version of the node image
+# This is a more minimal version of the node image.
 #
 # IMAGE
 # -----
-# * <registry>/infrastructure/node-mdw:x.x.x
+# * <registry>/infrastructure/node-lite:x.x.x
 #
 
 # python:3.10-slim-bookworm
@@ -25,12 +25,10 @@ WORKDIR /vantage6
 COPY README.md /vantage6/README.md
 COPY vantage6-common /vantage6/vantage6-common
 COPY vantage6-client /vantage6/vantage6-client
-COPY vantage6 /vantage6/vantage6
 COPY vantage6-node /vantage6/vantage6-node
 
 RUN pip install --upgrade pip \
     && pip install \
         -e /vantage6/vantage6-common \
         -e /vantage6/vantage6-client \
-        -e /vantage6/vantage6 \
         -e /vantage6/vantage6-node
